requirements.txt
torch
torchvision
torchaudio
Pillow